+++ packages/go-wasm/src/wasm_exec.d.ts
@@ -1,0 +1,2 @@
+declare function WasmExec(): void;
+export default WasmExec;